Adds doctest support to core
Adds a doctest job to the CI pipeline.

Removes the explicit doctest spec and instead uses a cabal repl with doctest to execute the tests.

This change adds the doctest tool to the nix shell and provides helper functions to run the doctests.

It also adds an example doctest to the Array module.